python/pyproject.toml
[project]
name = "argus-ml"
version = "0.1.0"
description = "Argus face recognition ML service (RetinaFace + ArcFace + ByteTrack)"
requires-python = ">=3.11,<3.12"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "numpy>=1.26,<2.0",
    "opencv-python-headless>=4.10",
    "onnxruntime>=1.20",
    "insightface>=0.7.3",
    "redis>=5.2",
    "pillow>=11.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.8",
    "mypy>=1.13",
]
eval = [
    "matplotlib>=3.9",
    "scikit-learn>=1.5",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["argus_ml*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
